Dowhistle MCP Server/dowhistle-mcp-server/pyproject.toml
[project]
name = "dowhistle-mcp-server"
version = "0.1.0"
description = "MCP Server with Agent Capabilities"
readme = "README.md"
requires-python = ">=3.13"
keywords = ["mcp", "server", "agents", "api","fastmcp","dowhistle"]
authors = [
    { name = "Nishanth M", email = "developernishanth2000@gmail.com" }
]
dependencies = [
    "fastmcp>=2.11.3",
    "uvicorn[standard]>=0.35.0",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "tenacity>=9.1.2",
    "structlog>=25.4.0",
    "openai>=1.58.1"
]
